Enable recusive scanning of resources during unit tests outside of JAR
